+++ gDRplots/R/helpers-drawResponseOverview.R
@@ -189,8 +189,11 @@ prepare_extras <- function(metrics,
 #' @param curve_data a data.table containing coordinates for plotting predicted dose response curves
 #' @param var_y which variable to plot: GR value or Relative Viability
 #' @param range_x numeric vector of length 2, specifying range of X axis
-#' @param plot_width,plot_height numeric; dimensions of plot
-#' 
+#' @param plot_width numeric, width of the plot in pixels
+#' @param plot_height numeric, height of the plot in pixels
+#' @param ticks_x numeric vector, specifying the positions of ticks on the X axis
+#' @param ticks_y numeric vector, specifying the positions of ticks on the Y axis
+#'
 #' @examples
 #' se <- gDRutils::get_synthetic_data("small")[[1]][1:3, 4:8]
 #' dt <- gDRutils::convert_se_assay_to_custom_dt(se, "Metrics")
@@ -199,7 +202,7 @@ prepare_extras <- function(metrics,
 #'
 #' @return A plotly object with highlights. The plot contains only response curves.
 #'
-#' The app will tracks clicks and create a secondary plot with a more detailed view of the selected curves.
+#' The app will track clicks and create a secondary plot with a more detailed view of the selected curves.
 #'
 #' @keywords drawResponseOverview
 #' @export
@@ -208,11 +211,15 @@ plotly_response_curve_all <- function(curve_data,
                                       var_y, 
                                       range_x = c(1e-3, 50e+0),
                                       plot_width = 400L,
-                                      plot_height = 300L) {
+                                      plot_height = 300L,
+                                      ticks_x = 10 ^ (-10:10),
+                                      ticks_y = -2:4 / 2) {
   
   checkmate::assert_numeric(plot_width, lower = 1)
   checkmate::assert_numeric(plot_height, lower = 1)
   checkmate::assert_numeric(range_x, len = 2)
+  checkmate::assert_numeric(ticks_x, min.len = 1)
+  checkmate::assert_numeric(ticks_y, min.len = 1)
   
   # get prettified identifiers
   pidfs <- gDRutils::get_prettified_identifiers(simplify = TRUE)
@@ -250,7 +257,10 @@ plotly_response_curve_all <- function(curve_data,
   range_y <- switch(var_y,
                     "GR value" = c(-1, 1.1),
                     "Relative Viability" = c(0, 1.1))
+  
   # line properties
+  range_x[1] <- max(ticks_x[ticks_x < range_x[1]])
+  
   line_horizontal_top <- list(
     type = "line", line = list(width = 1, color = "#A6A6A6"), layer = "below",
     x0 = range_x[1], x1 = range_x[2], y0 = 1, y1 = 1)
@@ -263,10 +273,6 @@ plotly_response_curve_all <- function(curve_data,
   line_vertical <- list(
     type = "line", line = list(width = 1), layer = "below",
     x0 = 1, x1 = 1, y0 = range_y[1], y1 = range_y[2])
-  
-  # positions of axis ticks
-  ticks_x <- 10 ^ (-10:10)
-  ticks_y <- -2:4 / 2
   
   # drop data points that would not fit in the plotting area
   # show exception_data if all data has been dropped
@@ -349,7 +355,7 @@ plotly_response_curve_all <- function(curve_data,
 #'
 #' @section Recursive plotting:
 #' If the data set contains multiple co-drugs, this function is called recursively
-#' so that data for each co-drug is plotted on a separate panel. Panel are arranged vertically.
+#' so that data for each co-drug is plotted on a separate panel. Panels are arranged vertically.
 #'
 #' @param data a data.table with cell fitness data taken from a \code{SummarizedExperiment}
 #' @param var_y variable to plot on Y axis
@@ -357,8 +363,11 @@ plotly_response_curve_all <- function(curve_data,
 #' @param curves a data.table with drug response predicted from growth metrics
 #' @param range_x numeric vector of length 2, specifying range of X axis
 #' @param extras a list of extra features that comprise the fifth layer, see \code{Details}
-#' @param plot_width,plot_height numeric; dimensions of plot
-#' 
+#' @param plot_width numeric, width of the plot in pixels
+#' @param plot_height numeric, height of the plot in pixels
+#' @param ticks_x numeric vector, specifying the positions of ticks on the X axis
+#' @param ticks_y numeric vector, specifying the positions of ticks on the Y axis
+#'
 #' @examples
 #' se <- gDRutils::get_synthetic_data("small")[[1]][1:3, 4:6]
 #' dt <- gDRutils::convert_se_assay_to_custom_dt(se, "Metrics")
@@ -387,11 +396,15 @@ plotly_response_curve_selected <- function(data,
                                            range_x = c(1e-3, 50e+0),
                                            extras,
                                            plot_width = 400L,
-                                           plot_height = 300L) {
+                                           plot_height = 300L,
+                                           ticks_x = 10 ^ (-10:10),
+                                           ticks_y = -2:4 / 2) {
   
   checkmate::assert_numeric(plot_width, lower = 1)
   checkmate::assert_numeric(plot_height, lower = 1)
   checkmate::assert_numeric(range_x, len = 2)
+  checkmate::assert_numeric(ticks_x, min.len = 1)
+  checkmate::assert_numeric(ticks_y, min.len = 1)
   
   # get prettified identifiers
   pidfs <- gDRutils::get_prettified_identifiers(simplify = TRUE)
@@ -492,6 +505,11 @@ plotly_response_curve_selected <- function(data,
   
   # line properties
   lines <- NULL # reset
+
+  
+  # line properties
+  range_x[1] <- max(ticks_x[ticks_x < range_x[1]])
+  
   line_horizontal_top <- list(
     type = "line", line = list(width = 1, color = "#A6A6A6"), layer = "below",
     x0 = range_x[1], x1 = range_x[2], y0 = 1, y1 = 1)
@@ -511,10 +529,6 @@ plotly_response_curve_selected <- function(data,
     names(lines_extra) <- NULL
     lines <- c(lines, lines_extra)
   }
-  
-  # positions of axis ticks
-  ticks_x <- 10 ^ (-10:10)
-  ticks_y <- -2:4 / 2
   # axis options
   axis_options_x <- list(title = sprintf("%s [&mu;M]", concentration_name), type = "log",
                          showgrid = FALSE, zeroline = FALSE, range = log10(range_x),