+++ gDRplots/NEWS.md
@@ -1,5 +1,8 @@
-## gDRplots 0.0.20 - 2024-08-02
+## gDRplots 0.0.21 - 2024-08-05
 * update plots for quality control of the distribution of normalized combo data
+
+## gDRplots 0.0.20 - 2024-08-05
+* extend the range of x-axis in dose-response curve
 
 ## gDRplots 0.0.19 - 2024-08-01
 * prevent drugCombo plot from overwriting the source data